Added functionality to classes
In this commit, we have added functionality to our classes and
utilized the data structures hashtable and set.
We have started implementing the methods for each one of these
entities.

NOTE: There is a compiler error here, very likely caused by the
codependent classes Transition and State. This should be fixed
in the next commit